session-04-telemetry/requirements.txt
gnmi-py
protobuf
python-dotenv
influxdb-client
grpcio